requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.103
botocore==1.34.103
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.1
dataclasses-json==0.5.14
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.9
django-allauth==0.60.1
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.4
django-celery-beat==2.6.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-sslserver==0.22
django-timezone-field==6.1.0
facebook-sdk==3.1.0
frozenlist==1.4.1
github==1.2.7
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.5
linkedin_api==2.1.0
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pillow==10.3.0
pipdeptree==2.16.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pygit2==1.14.1
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-facebook-api==0.19.0
python-twitter-v2==0.9.0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
